Simplified plugin's init by removing the inversion in the usage of toHide/toShow variables when settings.type=='random' Because of this inversion the code was confusing, and was actually showing items designated by *both* toHide and toShow in the random case.
Also, because toShow/toHide are determined depending on settings.type, I removed the line that was hiding all but the first element of $fade_object.
All elements are hidden instead in the z-index loop, which is more efficient (one loop on all elements instead of two) and sufficient as the toShow element will be shown right after the loop anyway.